SS-130: Adding functionality to display images on the profile window and profile list page, seprating the identification and role tabs into individual components, code cleanup.
