+++ web/src/components/MyTeam/MyTeamMembers.vue
@@ -23,11 +23,11 @@
       
 
         <div v-else class="container mb-5" id="app">
-            <div class="row">
+            <div class="row" :key="photokey">
                 <div v-for="teamMember in myTeamData" :key="teamMember.badgeNumber" class="col-3  my-1">
                     <div  class="card h-100">
                         <div @click="openMemberDetails(teamMember.id)" class="card-body">
-                            <user-summary-template :userBadgeNumber="teamMember.badgeNumber" :userName="teamMember.fullName" :userRole="teamMember.rank" :userImage="teamMember.image" :editMode="false" />
+                            <user-summary-template v-on:photoChange="photoChanged" :user="teamMember" :editMode="false" />
                         </div>
                         <div class="card-footer bg-light border-light" >                                                
                             <expire-sheriff-profile :userID="teamMember.id" :userIsEnable="teamMember.isEnabled"/>                        
@@ -45,101 +45,33 @@
             <b-card v-if="isUserDataMounted" no-body>
                 <b-row>
                     <b-col cols="3">
-                        <user-summary-template :userBadgeNumber="user.badgeNumber" :userName="user.firstName + ' ' + user.lastName" :userRole="user.rank" :userImage="user.image" :editMode="editMode"/>
+                        <user-summary-template v-on:photoChange="photoChanged" :user="user" :editMode="editMode"/>
                     </b-col>
                     <b-col cols="9">
                         <b-card no-body>
                             <b-tabs card v-model="tabIndex">
                                 <b-tab title="Identification" >
-                                    
-                                    <b-form-group v-if="createMode"><label>IDIR User Name<span class="text-danger">*</span></label>
-                                        <b-form-input v-model="user.idirUserName" placeholder="Enter IDIR User Name" :state = "idirUserNameState?null:false"></b-form-input>
-                                    </b-form-group>
-                                    <h2 class="mx-1 mt-0"><b-badge v-if="duplicateIdir" variant="danger"> Duplicate IDIR</b-badge></h2>
-
-                                    <b-row class="mx-1"> 
-                                        <b-form-group class="mr-1" style="width: 20rem"><label>First Name<span class="text-danger">*</span></label>
-                                            <b-form-input v-model="user.firstName" placeholder="Enter First Name" :state = "firstNameState?null:false"></b-form-input>
-                                        </b-form-group>                                    
-                                        <b-form-group class="ml-1" style="width: 20rem"><label>Last Name<span class="text-danger">*</span></label>
-                                            <b-form-input v-model="user.lastName" placeholder="Enter Last Name" :state = "lastNameState?null:false"></b-form-input>
-                                        </b-form-group>
-                                    </b-row>
-
-                                    <b-row class="mx-1">   
-                                        <b-form-group class="mr-1" style="width: 10rem"><label>Gender<span class="text-danger">*</span></label>
-                                            <b-form-select v-model="user.gender" :options="genderOptions" :state = "selectedGenderState?null:false"></b-form-select>
-                                        </b-form-group>
-                                        <b-form-group class="ml-1" style="width: 30rem"><label>Email<span class="text-danger">*</span></label>
-                                            <b-form-input v-model="user.email" placeholder="Enter Email" :state = "emailState?null:false" type="email"></b-form-input>
-                                        </b-form-group>
-                                    </b-row>
-
-                                    <b-row class="mx-1">
-                                        <b-form-group class="mr-1" style="width: 20rem"><label>Badge Number<span class="text-danger">*</span></label>
-                                            <b-form-input v-model="user.badgeNumber" placeholder="Enter Badge Number" :state = "badgeNumberState?null:false"></b-form-input>
-                                        </b-form-group>                                            
-                                        <b-form-group class="ml-1" style="width: 15rem"><label>Rank<span class="text-danger">*</span></label>
-                                            <b-form-select v-model="user.rank" placeholder="Select Rank" :options="commonInfo.sheriffRankList" :state = "selectedRankState?null:false"></b-form-select>
-                                        </b-form-group>
-                                    </b-row>
-                                    <h2 class="mx-1 mt-0"><b-badge v-if="duplicateBadge" variant="danger"> Duplicate Badge</b-badge></h2>
-                                    
+                                    <identification-tab 
+                                        :runMethod="identificationTabMethods" 
+                                        v-on:showWarning="$bvModal.show('bv-modal-team-cancel-warning')" 
+                                        v-on:closeMemberDetails="closeWarningWindow()" 
+                                        v-on:profileUpdated="getSheriffs()" 
+                                        :originalUser="user"
+                                        :createMode="createMode" 
+                                        :editMode="editMode" />
                                 </b-tab>
 
                                 <b-tab title="Locations">                                    
                                 </b-tab>
+
                                 <b-tab title="Leaves">                                    
                                 </b-tab>
+
                                 <b-tab title="Training"> 
                                 </b-tab>
+
                                 <b-tab v-if="userIsAdmin & editMode" title="Roles" class="p-0">
-                                    
-                                    <b-card  style="height:400px;overflow: auto;" >                                        
-                                        <h2 class="mx-1 mt-0"><b-badge v-if="roleAssignError" variant="danger"> Role Assignment Unsuccessful <b-icon class="ml-3" icon = x-square-fill @click="roleAssignError = false" /></b-badge></h2>
-                                               
-                                        <b-form-group >
-                                            <b-form-checkbox-group v-model="selectedRoles">
-                                                <b-row class="mb-2 text-primary" style="font-weight:bold">
-                                                    <b-col cols="5"> Role </b-col>
-                                                    <b-col class="mx-2" > Effective Date </b-col>
-                                                    <b-col class="mx-2" > Expiry Date </b-col>
-                                                </b-row>
-                                                <b-row v-for="(rol,rolInx) in roles" :key="rolInx" class="mb-1">
-                                                    <b-col cols="5">   
-                                                        <b-form-checkbox 
-                                                            class="mt-1"
-                                                            v-b-tooltip.hover.left
-                                                            :title="rol.desc"
-                                                            @change="roleChanged(rolInx)" 
-                                                            :value="rol.value">
-                                                                {{rol.text}}
-                                                        </b-form-checkbox>
-                                                    </b-col>
-                                                    <b-col >
-                                                        <b-form-datepicker
-                                                        class="p-0 m-0"
-                                                        v-model="roles[rolInx].effDate"
-                                                        placeholder="Eff. Date"
-                                                        locale="en-US" 
-                                                        :date-format-options="{ year: 'numeric', month: 'short', day: '2-digit' }"
-                                                        :state = "rol.effState?null:false"
-                                                        ></b-form-datepicker>
-                                                    </b-col>
-                                                    <b-col >
-                                                        <b-form-datepicker
-                                                        v-model="roles[rolInx].expDate"
-                                                        placeholder="Exp. Date"
-                                                        locale="en-US"
-                                                        :date-format-options="{ year: 'numeric', month: 'short', day: '2-digit' }"
-                                                        :state = "rol.expState?null:false"
-                                                        ></b-form-datepicker>
-                                                    </b-col>
-                                                </b-row> 
-                                          
-                                            </b-form-checkbox-group>
-                                        </b-form-group>                                       
-                                    </b-card>
+                                    <role-assignment-tab :userId="user.id" :userActiveRoles="userActiveRoles" />
                                 </b-tab>
                                 
 
@@ -194,7 +126,7 @@
     import { namespace } from 'vuex-class';
     import * as _ from 'underscore';
     import PageHeader from "@components/common/PageHeader.vue";
-    import UserSummaryTemplate from "./UserSummaryTemplate.vue";
+    import UserSummaryTemplate from "./utils/UserSummaryTemplate.vue";
     import "@store/modules/CommonInformation";  
     import {commonInfoType, locationInfoType, userInfoType} from '../../types/common';
     import {teamMemberInfoType, roleOptionInfoType} from '../../types/MyTeam';
@@ -202,6 +134,8 @@
     const commonState = namespace("CommonInformation");
     import store from '../../store'
     import ExpireSheriffProfile from './utils/ExpireSheriffProfile.vue'
+    import RoleAssignmentTab from './utils/RoleAssignmentTab.vue'
+    import IdentificationTab from './utils/IdentificationTab.vue'
 
     enum gender {'Male'=0, 'Female', 'Other'}
 
@@ -209,7 +143,9 @@
         components: {
             PageHeader,
             UserSummaryTemplate,
-            ExpireSheriffProfile
+            ExpireSheriffProfile,
+            RoleAssignmentTab,
+            IdentificationTab
         }        
     })    
     export default class MyTeamMembers extends Vue {
@@ -220,7 +156,7 @@
         @commonState.Action
         public UpdateCommonInfo!: (newCommonInfo: commonInfoType) => void
 
-         @commonState.State
+        @commonState.State
         public token!: string;
 
         @commonState.Action
@@ -238,32 +174,16 @@
         showMemberDetails = false;
         showCancelWarning = false;
         user = {} as teamMemberInfoType;
-        originalUser = {} as teamMemberInfoType;
-        userJson;
-        genderOptions = [{text:"Male", value: gender.Male}, {text:"Female", value: gender.Female}, {text:"Other", value: gender.Other}]
-        genderValues = [0, 1, 2]
-        firstNameState = true;
-        lastNameState = true;
-        emailState = true;
-        selectedGenderState = true;
-        badgeNumberState = true;
-        selectedRankState = true;
-        idirUserNameState = true;
-        duplicateBadge = false;
-        duplicateIdir = false;
         userIsAdmin = false;
 
         tabIndex = 0;
-        errorCode = 0;
-        errorText = '';
         isUserDataMounted = false;
         editMode = false;
         createMode = false;
         sectionHeader = '';
-        selectedRoles: string[] = []
-        roles: roleOptionInfoType[] = []
-        roleAssignError = false;
 
+        photokey=0;
+        userActiveRoles: any[] = []
 
         isMyTeamDataMounted = false;
         myTeamData: teamMemberInfoType[] =[];
@@ -281,167 +201,77 @@
             this.sectionHeader = "My Team - " + this.location.name;
         }
 
-        public getSheriffs()
-        {
+        public getSheriffs(){
             this.isMyTeamDataMounted = false;
             const url = 'api/sheriff?locationId=' + this.location.id
             const options = {headers:{'Authorization' :'Bearer '+this.token}}
             this.$http.get(url, options)
                 .then(response => {
                     if(response.data){
-                        // console.log(response.data)
-                        this.extractMyTeam(response.data);                        
+                        console.log(response.data)
+                        this.extractMyTeamFromSheriffs(response.data);                        
                     }
                     this.isMyTeamDataMounted = true;
                 })
         }
 
-        public extractMyTeam(data: any)
-        {
+        public extractMyTeamFromSheriffs(data: any){
             this.myTeamData = [];            
             for(const myteaminfo of data)
             {
                 const myteam: teamMemberInfoType = {id:'',idirUserName:'', rank:'', firstName:'', lastName:'', email:'', badgeNumber:'', gender:'' }
-                myteam.fullName = this.getFullNameOfPerson(myteaminfo.firstName ,myteaminfo.lastName);
+                myteam.fullName = Vue.filter('capitilize')(myteaminfo.firstName) + ' ' + Vue.filter('capitilize')(myteaminfo.lastName);
                 myteam.rank = myteaminfo.rank;
                 myteam.badgeNumber = myteaminfo.badgeNumber;
                 myteam.id = myteaminfo.id;
+                myteam.image = myteaminfo.photo? 'data:image/;base64,'+myteaminfo.photo: '';
                 myteam.isEnabled = myteaminfo.isEnabled;
                 this.myTeamData.push(myteam);
             }
         }
 
-        public getFullNameOfPerson(first: string, last: string)
-        {            
-            return Vue.filter('capitilize')(first) + ' ' + Vue.filter('capitilize')(last);
-        }
+        
 
-        public openMemberDetails(userId)
-        {
+        public photoChanged(id: string, image: string){
+           
+            const index = this.myTeamData.findIndex(myteam => {if(myteam.id == id) return true;else return false})
+            if( index >=0 ){
+                this.myTeamData[index].image = image;
+                this.photokey++;
+            }
+        }      
+
+        public openMemberDetails(userId){
             this.createMode = false;
-            this.editMode = true;
-            this.badgeNumberState = true;
-            this.idirUserNameState = true;
-            this.duplicateBadge = false;
-            this.duplicateIdir = false;
-            this.GetRoles(userId);      
-            
+            this.editMode = true;            
+            this.loadUserDetails(userId);            
         }
 
-        public closeProfileWindow() 
-        {         
-            if(this.createMode && this.isEmpty(this.user))
-            {
-                this.showMemberDetails = false;
-                this.resetProfileWindowState();
-            } 
-            else if(this.editMode && !this.changesMade())
-            {
-                this.showMemberDetails = false;
-                this.resetProfileWindowState();
-            }    
-            else
-                this.showCancelWarning = true;
+        identificationTabMethods = new Vue();
+
+        public closeProfileWindow(){
+            this.identificationTabMethods.$emit('closeProfileWindow'); 
         }
 
-        public changesMade(): boolean {
-            return !_.isEqual(this.originalUser, this.user)
-        }
 
-        public isEmpty(obj){
-            for(const prop in obj) 
-                if(obj[prop] != null)
-                    return false;
-            return true;
-        }
-
-        public closeWarningWindow() {   
+        public closeWarningWindow() {             
             this.resetProfileWindowState();         
             this.showCancelWarning = false;
-            this.showMemberDetails = false;
+            this.showMemberDetails = false;            
         }
 
         public resetProfileWindowState() {
             this.createMode = false;
             this.editMode = false;
-            this.firstNameState = true;
-            this.lastNameState = true;
-            this.selectedGenderState = true;
-            this.badgeNumberState = true;
-            this.emailState = true;
-            this.selectedRankState = true;
-            this.idirUserNameState = true;
-            this.duplicateBadge = false;
-            this.duplicateIdir = false;
             this.user = {} as teamMemberInfoType;
         }
 
-        public AddMember()
-        {  
+
+        public AddMember(){ 
             this.createMode = true;
             this.editMode = false;
             this.isUserDataMounted = true;
-            this.showMemberDetails=true;
-        }
-
-        public GetRoles(userId){
-            const url = '/api/role'
-            const options = {headers:{'Authorization' :'Bearer '+this.token}}
-            this.$http.get(url, options)
-                .then(response => {
-                    if(response.data){
-                        this.extractRoles(response.data);                        
-                    }
-                    this.loadUserDetails(userId);                    
-                })
-        }
-
-        public roleChanged(rolInx){
-            Vue.nextTick().then(()=>{
-                const selectedIndex = this.selectedRoles.indexOf(this.roles[rolInx].value);
-                console.log(rolInx)
-                console.log(this.roles[rolInx].value)
-                this.roles[rolInx].effState = true;
-                this.roles[rolInx].expState = true; 
-                this.roleAssignError = false; 
-
-                if(selectedIndex >=0 && this.roles[rolInx].effDate=="")
-                {
-                   this.roles[rolInx].effState = false; 
-                   this.selectedRoles.splice(selectedIndex, 1);
-                }
-                else 
-                {
-                    const body = 
-                    [{
-                        "userId": this.user.id,
-                        "roleId": this.roles[rolInx].value,
-                        "effectiveDate": this.roles[rolInx].effDate,
-                        "expiryDate": this.roles[rolInx].expDate
-                    }]
-                    const url = (selectedIndex >= 0)? 'api/sheriff/assignroles' :'api/sheriff/unassignroles' 
-                    const options = {headers:{'Authorization' :'Bearer '+this.token}}
-                    this.$http.put(url, body, options)
-                        .then(response => {
-                            console.log(response)
-                            console.log('assign success')  
-                            if(selectedIndex < 0)
-                            {
-                                this.roles[rolInx].effDate =""
-                                this.roles[rolInx].expDate="" 
-                            }                                              
-                        }, err=>{this.roleAssignError = true;});
-                }
-            });
-        }
-
-        public extractRoles(data){
-            this.roles=[];
-            this.selectedRoles = [];
-            this.roleAssignError = false; 
-            
-            for(const role of data)            
-                this.roles.push({text:role.name, desc: role.description, value:role.id, effDate:'', expDate:'', effState:true, expState: true})           
+            this.showMemberDetails = true;
         }
 
         public loadUserDetails(userId): void {
@@ -451,181 +281,39 @@
             this.$http.get(url, options)
                 .then(response => {
                     if(response.data){
-                        this.userJson = response.data;
-                        this.extractUserInfo();
+                        console.log(response.data)                        
+                        this.extractUserInfo(response.data);
                         this.isUserDataMounted = true;
-                        this.showMemberDetails=true; 
-                        console.log(response.data)                       
+                        this.showMemberDetails=true;                                              
                     }                    
                 });
         }
 
-        public extractUserInfo(): void {
+        public extractUserInfo(userJson): void {
+            console.log(userJson)
             this.user = {} as teamMemberInfoType;            
-            this.user.idirUserName = this.originalUser.idirUserName = this.userJson.idirName;
-            this.user.firstName = this.originalUser.firstName = this.userJson.firstName;
-            this.user.lastName = this.originalUser.lastName = this.userJson.lastName;
-            this.user.gender = this.originalUser.gender = gender[this.userJson.gender];
-            this.user.rank = this.originalUser.rank = this.userJson.rank;
-            this.user.email = this.originalUser.email = this.userJson.email;
-            this.user.badgeNumber = this.originalUser.badgeNumber = this.userJson.badgeNumber;
-            this.user.id = this.originalUser.id = this.userJson.id;
-            this.user.image = this.originalUser.image = this.userJson['image']? this.userJson['image'] :null ; 
-            for(const activeRole of this.userJson.activeRoles) 
-            {             
-                const index = this.roles.findIndex(role =>{if(role.value == activeRole.role.id) return true;else return false});
-                if(index >=0)
-                { 
-                    this.roles[index].effDate = activeRole.effectiveDate
-                    this.roles[index].expDate = activeRole.expiryDate
-                    this.selectedRoles.push(this.roles[index].value);
-                }
-            }
+            this.user.idirUserName =  userJson.idirName;
+            this.user.firstName = userJson.firstName;
+            this.user.lastName = userJson.lastName;
+            this.user.fullName = Vue.filter('capitilize')(userJson.firstName) + ' ' + Vue.filter('capitilize')(userJson.lastName);
+            this.user.gender = gender[userJson.gender];
+            this.user.rank = userJson.rank;
+            this.user.email = userJson.email;
+            this.user.badgeNumber = userJson.badgeNumber;
+            this.user.id = userJson.id;
+            this.user.image = userJson['photo']?'data:image/;base64,'+userJson['photo']:''; 
+            console.log(this.user)
+            this.userActiveRoles = userJson.activeRoles
         }
 
         get getCancelLabel(){
             if(this.tabIndex<1) return 'Cancel'; else return 'Close'
         }
 
-        public saveMemberProfile() {       
-            const requiredErrorTab: number[] = [];
+        public saveMemberProfile() { 
+            this.identificationTabMethods.$emit('saveMemberProfile');
+        }      
 
-            if (this.createMode && !this.user.idirUserName) {
-                this.idirUserNameState = false;
-                requiredErrorTab.push(0);
-            } else {
-                this.idirUserNameState = true;
-                this.duplicateIdir = false;
-            }
-            if (!this.user.firstName) {
-                this.firstNameState = false;
-                requiredErrorTab.push(0);
-            } else {
-                this.firstNameState = true;
-            }
-            if (!this.user.lastName) {
-                this.lastNameState = false;
-                requiredErrorTab.push(0);
-            } else {
-                this.lastNameState = true;
-            }
-            if (this.genderValues.toString().indexOf(this.user.gender) == -1) {
-                this.selectedGenderState = false;
-                requiredErrorTab.push(0);
-            } else {
-                this.selectedGenderState = true;
-            }
-            if (!this.user.badgeNumber) {
-                this.badgeNumberState = false;
-                requiredErrorTab.push(0);
-            } else {
-                this.badgeNumberState = true;
-                this.duplicateBadge = false;
-            }
-            if (!this.user.email) {
-                this.emailState = false;
-                requiredErrorTab.push(0);
-            } else {
-                this.emailState = true;
-            }
-            if (!this.user.rank) {
-                this.selectedRankState = false;
-                requiredErrorTab.push(0);
-            } else {
-                this.selectedRankState = true;
-            }
-            
-            if (requiredErrorTab.length == 0) {
-                if (this.editMode) this.updateProfile();
-                if (this.createMode) this.createProfile();              
-
-            } else {                
-                this.tabIndex= requiredErrorTab[0];
-            }             
-        }
-
-        public updateProfile(): void {
-            const body = {
-                homeLocationId: this.location.id,               
-                gender: this.user.gender,
-                badgeNumber: this.user.badgeNumber,
-                rank: this.user.rank,
-                idirName: this.user.idirUserName,
-                firstName:this.user.firstName,
-                lastName: this.user.lastName,
-                email: this.user.email,
-                id: this.user.id
-            }
-
-            const url = 'api/sheriff';
-            const options = {headers:{'Authorization' :'Bearer '+this.token}} 
-            this.$http.put(url, body, options)
-                .then(response => {
-                    if(response.data){
-                        this.resetProfileWindowState();
-                        this.showMemberDetails = false;
-                        this.getSheriffs();                     
-                    }                    
-                }, err => {
-                    console.log(err)
-                    this.errorText = err.response.data.error
-                    this.errorCode = err.response.status
-                     
-                    //if(this.errorText.includes('already has badge number'))
-                    // has IDIR name
-                    if(err.response.status == 400)
-                    {
-                        if (this.errorText.includes('already has badge number')){
-                            this.badgeNumberState = false;
-                            this.duplicateBadge = true;
-                        } else if (this.errorText.includes('has IDIR name')) {
-                            this.idirUserNameState = false;
-                            this.duplicateIdir = true;
-                        }                        
-                    }
-
-                });
-        }
-
-        public createProfile() {
-            const body = {
-                homeLocationId: this.location.id,               
-                gender: this.user.gender,
-                badgeNumber: this.user.badgeNumber,
-                rank: this.user.rank,
-                idirName: this.user.idirUserName,
-                firstName:this.user.firstName,
-                lastName: this.user.lastName,
-                email: this.user.email
-            }
-            // console.log(body)
-            const url = 'api/sheriff';
-            const options = {headers:{'Authorization' :'Bearer '+this.token}}           
-            
-            this.$http.post(url, body, options )
-                .then(response => {
-                    if(response.data){
-                        this.resetProfileWindowState();
-                        this.showMemberDetails = false;
-                        this.getSheriffs();                     
-                    }
-                }, err => {
-                    this.errorText = err.response.data.error
-                    this.errorCode = err.response.status                     
-                    
-                    if(err.response.status == 400)
-                    {
-                        if (this.errorText.includes('already has badge number')){
-                            this.badgeNumberState = false;
-                            this.duplicateBadge = true;
-                        } else if (this.errorText.includes('has IDIR name')) {
-                            this.idirUserNameState = false;
-                            this.duplicateIdir = true;
-                        }                        
-                    }
-
-                })   
-        }
 
     }
 </script>
